pyproject.toml
[project]
name = "cats-api"
version = "0.1.0"
description = "FastAPI application for managing cats and users with MongoDB"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "motor==3.3.2",
    "pymongo==4.6.0",
    "pydantic==2.5.0",
    "bcrypt==4.1.2",
    "python-jose[cryptography]==3.3.0",
    "httpx==0.25.2",
    "python-multipart==0.0.6",
    "python-dotenv==1.0.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-mock==3.12.0"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --tb=short"
testpaths = ["tests"]
asyncio_mode = "auto"
